refactor(Wallpaper): move salon files into respective component dirs
- Move each salon sub-dir/file from centralized salon/ into corresponding component's salon/
- Rename TextureSwatchPreview -> SwatchPreview with matching salon
- Extract swatch layout constants into TextureStylePicker/constant.ts
- Delete unused salon/diy_tab.ts
- Rename texture_style_picker.ts -> index.ts in TextureStylePicker/salon/
