+++ frontend/core/unit/DashboardThread/Appearance/Wallpaper/GradientTab/salon/index.ts
