+++ frontend/core/unit/DashboardThread/Appearance/Wallpaper/GradientTab/salon/pattern_cards.ts
