+++ frontend/core/unit/DashboardThread/Appearance/Wallpaper/PicturesTab/salon/index.ts
