+++ frontend/core/unit/DashboardThread/Appearance/Wallpaper/PreviewPanel/salon/auth_preview.ts
