+++ frontend/core/unit/DashboardThread/Appearance/Wallpaper/PreviewPanel/salon/global_preview.ts
