+++ frontend/core/unit/DashboardThread/Appearance/Wallpaper/PreviewPanel/salon/index.ts
