+++ frontend/core/unit/DashboardThread/Appearance/Wallpaper/TextureStylePicker/salon/swatch_preview.ts
