+++ frontend/core/unit/DashboardThread/Appearance/Wallpaper/TuningPanel/DetailPanel/Gradient/index.tsx
@@ -15,8 +15,8 @@ import ColorSelector from '~/widgets/ColorSelector'
 import FocalPointControl, { type TPoint } from '~/widgets/FocalPointControl'
 import RangeInput from '~/widgets/RangeInput'
 
-import useSalon from '../../../salon/tuning_panel/detail_panel/gradient'
 import useLogic from '../../../useLogic'
+import useSalon from '../../salon/detail_panel/gradient'
 import AngleWheel from '../AngleWheel'
 import GroupItem from '../GroupItem'
 import GroupTitle from '../GroupTitle'
