+++ frontend/core/unit/DashboardThread/Appearance/Wallpaper/TuningPanel/salon/detail_panel/group_title.ts
