+++ frontend/core/unit/DashboardThread/Appearance/Wallpaper/TuningPanel/salon/detail_panel/texture.ts
