+++ frontend/core/unit/DashboardThread/Appearance/Wallpaper/TuningPanel/salon/hud_panel.ts
