+++ frontend/core/unit/DashboardThread/Appearance/Wallpaper/TuningPanel/salon/index.ts
