+++ frontend/core/unit/DashboardThread/Appearance/Wallpaper/UploadTab/salon/index.ts
