+++ frontend/core/unit/DashboardThread/Appearance/Wallpaper/UploadTab/salon/upload_box.ts
