--- frontend/core/unit/DashboardThread/Appearance/Wallpaper/salon/diy_tab.ts
@@ -1,25 +1,0 @@
-import useTwBelt from '~/hooks/useTwBelt'
-
-export { cn } from '~/css'
-
-export default function useSalon() {
-  const { cn, br, primary, fill } = useTwBelt()
-
-  return {
-    wrapper: 'column gap-5 mt-2.5',
-    presets: 'row-center wrap gap-3',
-    presetCard: cn(
-      'size-14 rounded-lg overflow-hidden relative border border-2 border-transparent pointer trans-all-200',
-      br('divider'),
-      `hover:${br('digest')}`,
-    ),
-    presetActive: br('digest'),
-    presetPreview: 's-full',
-    activeSign: cn(
-      'size-5 circle absolute -top-1 -right-0.5 z-20 border',
-      primary('bg'),
-      br('title'),
-    ),
-    checkIcon: cn('size-3.5 absolute top-0.5 left-0.5', fill('button.fg')),
-  }
-}
